Move common functions to ide-common.lua
